Changed route to /event/edit Moved to route to routers Moved event test to test_event.py
