+++ app/dependencies.py
@@ -1,0 +1,5 @@
+import os
+
+APP_PATH = os.path.dirname(os.path.realpath(__file__))
+STATIC_PATH = os.path.join(APP_PATH, "static")
+TEMPLATES_PATH = os.path.join(APP_PATH, "templates")
