+++ app/templates/eventedit.html
@@ -25,7 +25,7 @@
             </ul>
             <div class="tab-content" id="event_edit_tabs">
                 <div class="tab-pane fade show active" id="eventdetails" role="tabpanel" aria-labelledby="eventdetails-tab">
-                    {% include "eventdetailsformtab.html" %}
+                    {% include "event_details_form_tab.html" %}
                 </div>
                 <!--    Copy commented section to add another tab-->
                 <!--    <div class="tab-pane fade" id="(CHANGE_ME)" role="tabpanel" aria-labelledby="(CHANGE_ME)-tab">-->
