--- tests/test_app.py
@@ -1,9 +1,0 @@
-from fastapi.testclient import TestClient
-from app.main import app
-
-client = TestClient(app)
-
-def test_eventedit():
-    response = client.get("/eventedit")
-    assert response.status_code == 200
-    assert b"Edit Event" in response.content
